0603 1337 goosebeomsu: StorageMapper 수정

+++ Bit-forRest/src/main/java/com/mvc/forrest/service/domain/Storage.java
@@ -21,7 +21,7 @@ public class Storage {
 	private Date startDate;
 	private Date endDate;
 	private int period;
-	private short tranCode;
+	private int tranCode;
 	private String paymentNo;
 	private Timestamp  paymentDate;
 	private String paymentWay;
@@ -31,8 +31,6 @@ public class Storage {
 	private String prodImg;
 	private int originPrice;
 	private double discount ;
-	private int resultPrice ;
-	
-
+	private int resultPrice;
 	
 }
